requirements.txt
flask_mysqldb
flask
flask_json
RPi.GPIO
flask-security
gunicorn
Werkzeug
ptvsd